jamal

+++ 3NWSTb/2023-2024/INFORMATICA/3/Cirkel.py
@@ -1,5 +1,5 @@
 from math import *
 
-straal = int(input('straal?   '))
+straal = float(input('straal?   '))
 oppervlaakte = round(straal*straal*pi, 2)
 print(str(oppervlaakte) + "cm²")
