+++ 3NWSTb/2023-2024/INFORMATICA/3/Datatype.py
@@ -3,7 +3,7 @@ var1 = input("tekst getal: ")
 print("Dubbel van tekst", var1*2)
 
 #getal inlezen als een integer
-var2 = int(input("geheel getal: "))
+var2 = float(input("geheel getal: "))
 print("Dubbel van een getal", var2*2)
 
 print(type(var1))
