+++ 3NWSTb/2023-2024/INFORMATICA/3/Kubus.py
@@ -1,4 +1,4 @@
-ribbe = int(input("lengte ribbe:  "))
+ribbe = float(input("lengte ribbe:  "))
 
 oppervlakte = str(ribbe**2*6)+"cm²"
 volume = str(ribbe**3)
