+++ 3NWSTb/2023-2024/INFORMATICA/3/Pythagoras.py
@@ -1,6 +1,6 @@
 from math import *
 
-zijde1 = int(input("eerste zijde:   "))
-zijde2 = int(input("tweede zijde:   "))
+zijde1 = float(input("eerste zijde:   "))
+zijde2 = float(input("tweede zijde:   "))
 
 print(sqrt((zijde1**2)+(zijde2**2)))
